Create Create Machine learning project using python
The aim of the project is to predict six Erythemato-squamous diseases namely psoriasis(1,) seborrheic dermatitis(2), lichen planus(3), pityriasis rosea(4), chronic dermatitis(5), and pityriasis rubra pilaris(6). Traditional diagnostic methods, such as clinical evaluation and histological examination, often necessitate a biopsy to confirm a diagnosis. This study delves into the utilization of machine learning models and classifiers to identify the type of skin disease based on 11 clinical and 22 histopathological features. The project aims to construct five models, which include a regression model to identify the disease type based on the patient's age, a random forest and k-nearest neighbors (KNN) to classify the disease type and two clustering algorithms.